+++ simpleAdmin/admin/userDelete.php
@@ -1,7 +1,15 @@
 <?php
-include_once "includes/config.inc.php";
-$con = mysqli_connect(DBCON,DBUSER,DBPW,DBNAME);
-$sql = "DELETE FROM users WHERE userId = ".$_GET['id'];
-mysqli_query($con,$sql);
-header("location: users.php");
+include "includes/config.inc.php";
+include "includes/functions.inc.php";
+session_start();
+$state = validateSessions($_SESSION['userId']);
+
+if ($state == false){
+    header("location: login.php?err");
+}else {
+    $con = mysqli_connect(DBCON, DBUSER, DBPW, DBNAME);
+    $sql = "DELETE FROM users WHERE userId = " . $_GET['id'];
+    mysqli_query($con, $sql);
+    header("location: users.php");
+}
 ?>
